requirements.txt
streamlit
Pillow
opencv-python-headless
numpy